# Make loading IP lists exponentially faster