@robertdavidgraham 2021-05-26
This has been fixed for a while. It should load 10-million addresses pretty quickly